Review modifications & fix bug & update Go client version
