+++ node/rpc/src/config.rs
@@ -5,4 +5,5 @@ pub struct Config {
     pub enabled: bool,
     pub listen_address: SocketAddr,
     pub chunks_per_segment: usize,
+    pub max_cache_file_size: usize,
 }
